Install: Allow override of files that have no owner
In a package install, the function check_data_file_clashes()
checks if a file that is going to be installed exists
and if it belongs to any package.

Currently if a file exists, the installation fails as
a result of a file clash.

This patch changes a bit that behavior by allowing the
overwrite of files that have no owner, i.e., don't
belong to any package.

This is, for instance, the behavior of apt.

Signed-off-by: Jonh Wendell <jonh.wendell@oiwifi.com.br>